+++ CHANGELOG.md
@@ -3,6 +3,12 @@ This file keep track of the changes.
 
 ## UNRELEASED
 
+# [1.2.0] - 2023-12-18
+### CHANGED
+- Require at least php 8.3
+- Use the new symfony mailer library
+- Update Symfony 5.4 to 6.4
+
 # [1.1.0] - 2023-12-18
 ### ADDED
 - Ajout des templates manquants